+++ src/Cache/Configuration/Configuration.php
@@ -12,7 +12,7 @@ final class Configuration implements \Countable
      * @param array<non-empty-string, list<array{
      *     routeName: string,
      *     routeParams?: array<string, array{type: string, values: list<mixed>, optional?: true}>,
-     *     if?: string,
+     *     if?: string|array<mixed>,
      *     actions?: non-empty-list<Action>,
      * }>> $configuration
      */
@@ -56,7 +56,7 @@ final class Configuration implements \Countable
      * @return array<non-empty-string, list<array{
      *     routeName: string,
      *     routeParams?: array<string, array{type: string, values: list<mixed>, optional?: true}>,
-     *     if?: string,
+     *     if?: string|array<mixed>,
      *     actions?: non-empty-list<Action>,
      * }>>
      */