+++ src/Cache/Configuration/ConfigurationLoader.php
@@ -38,7 +38,11 @@ final class ConfigurationLoader implements ConfigurationLoaderInterface
             }
 
             if (null !== $subscription->if) {
-                $config['if'] = (string) $subscription->if;
+                if ($subscription->if instanceof \Closure) {
+                    $config['if'] = deepclone_to_array($subscription->if);
+                } else {
+                    $config['if'] = (string) $subscription->if;
+                }
             }
 
             if (null !== $subscription->actions) {