+++ src/Cache/PropertyResolver/AssociationResolver.php
@@ -73,6 +73,11 @@ final class AssociationResolver implements SubscriptionResolverInterface
         }
 
         if (null !== $if = $routeMetadata->purgeOn->if) {
+            if ($if instanceof \Closure) {
+                // TODO support closures
+                throw new \RuntimeException('Cannot create inverse subscription with closures');
+            }
+
             $if = $this->expressionTransformer->transform($if, $associationClass, $associationTarget, 'false');
         }